Add more startup display capability fix README

+++ startup.sh
@@ -19,8 +19,12 @@ echo "My address is $MY_ADDRESS"
 sleep 2
 ./showtext 'quit quit quit to exit'
 
+touch /tmp/piday.out
 python ./example_recognizer.py 2>/tmp/piday.err >/tmp/piday.out
+rc=$?
 sleep 2
-echo 'shutting down'
-./showtext 'shutting down'
-sudo poweroff
+if [[ $rc -eq 0 ]]; then
+  echo 'shutting down'
+  ./showtext 'shutting down'
+  sudo poweroff
+fi
